refactor: extract source group actions
